+++ metax/taxafunc_analyzer/analyzer.py
@@ -54,6 +54,7 @@ class TaxaFuncAnalyzer:
 
         self.peptide_col_name = peptide_col_name
         self.protein_col_name = protein_col_name
+        self.protein_separator = ';'
         self.custom_col_name = custom_col_name
         self.sample_list: Optional[List[str]] = None
         self.meta_df: Optional[pd.DataFrame] = None
@@ -78,6 +79,7 @@ class TaxaFuncAnalyzer:
         self.any_df_mode = any_df_mode  # if True, the consider the TaxaFunc df as other_df
         self.custom_df: Optional[pd.DataFrame] = None # other df, any df that user want to add
         self.peptide_num_used = {'taxa': 0, 'func': 0, 'taxa_func': 0, 'protein': 0}
+        self.distinct_peptides_list: list|None = None
         
         self.split_func_status:bool = False
         self.split_func_sep:str = ''
@@ -689,6 +691,18 @@ class TaxaFuncAnalyzer:
         
         return df_taxa_func
     
+    def calculate_distinct_peptides(self): #! NOT USED YET
+        # extract the peptide column and protein_col_name
+        print("Calculating distinct peptides list...")
+        extract_cols = [self.peptide_col_name, self.protein_col_name]
+        df = self.original_df[extract_cols]
+        separate_protein = self.protein_separator
+        df['protein_num'] = df[self.protein_col_name].apply(lambda x: len(x.split(separate_protein)))
+        df = df[df['protein_num'] == 1]
+        distinct_peptides = df[self.peptide_col_name].tolist()
+        self.distinct_peptides_list = distinct_peptides
+        
+    
     def update_data_preprocess_parameters(self, data_preprocess_params):
         
         normalize_method = data_preprocess_params['normalize_method']
@@ -706,13 +720,12 @@ class TaxaFuncAnalyzer:
         
         return data_preprocess_params
 
-    def filter_peptides_num_for_splited_func(self, df, peptide_num_threshold, df_type, distinct_threshold_mode=False):
+    def filter_peptides_num_for_splited_func(self, df, peptide_num_threshold, df_type):
         '''
         Only for the splited func table or taxa_func table
         - df: the splited func table or taxa_func table which has been grouped, index is the func or taxa_func
         - peptide_num_threshold: the threshold of peptide number for each func or taxa_func
         - df_type: 'func' or 'taxa_func'
-        - distinct_threshold_mode: TODO
         '''
         
         valid_df_types = ['func', 'taxa_func']
@@ -750,6 +763,49 @@ class TaxaFuncAnalyzer:
         else:
             item_col = 'Taxon' if df_type == 'taxa' else self.func_name
 
+        # # if True: #! Need to be implemented
+        # if distinct_threshold_mode:
+        #     if self.distinct_peptides_list is None:
+        #         self.calculate_distinct_peptides()
+            
+        #     peptides_in_taxa_func = defaultdict(list)
+        #     peptides_in_taxa = defaultdict(list)
+        #     peptides_in_func = defaultdict(list)
+        #     skiped_peptides_list = []
+        #     for row in tqdm(df.itertuples(index=False), total=len(df), desc="Creating peptides_dict"):
+        #         peptide = row[0]
+        #         if peptide not in self.distinct_peptides_list:
+        #             skiped_peptides_list.append(peptide)
+        #             continue
+                
+        #         if df_type == 'taxa':
+        #             taxa = row[1]
+        #             # Append peptide to taxa list
+        #             peptides_in_taxa[taxa].append(peptide)
+                    
+        #         if self.split_func_status:
+        #                 func_list = [f.strip() for f in row[2].split(self.split_func_sep)]
+        #                 # Process each function in the func_list
+        #                 for func in func_list:
+        #                     peptides_in_func[func].append(peptide)
+        #                     taxa_func = f'{taxa}&&&&{func}'
+        #                     peptides_in_taxa_func[taxa_func].append(peptide)
+        #         else:
+        #             if df_type in ['func', 'taxa_func']:
+        #                 taxa = row[1]
+        #                 func = row[2]
+        #                 # Append peptide to func list
+        #                 peptides_in_func[func].append(peptide)
+        #                 # Create combined key for taxa_func
+        #                 taxa_func = f'{taxa}&&&&{func}'
+        #                 peptides_in_taxa_func[taxa_func].append(peptide)
+
+        #     peitides_dict = {'taxa': peptides_in_taxa, 'func': peptides_in_func, 'taxa_func': peptides_in_taxa_func}
+        #     remove_list = [k for k, v in peitides_dict[df_type].items() if len(v) < peptide_num]
+        #     skiped_peptides_list = set(skiped_peptides_list)
+
+
+        # else:                
         # Group by item_col and filter based on peptide number
         dict_item_pep_num = df.groupby(item_col).size().to_dict()
         remove_list = [k for k, v in dict_item_pep_num.items() if v < peptide_num]
@@ -761,7 +817,7 @@ class TaxaFuncAnalyzer:
             df = df.drop('taxa_func', axis=1)
 
         self.peptide_num_used[df_type] = len(df)
-        print(f"Removed [{len(remove_list)} {df_type}] from [{df_original_len - len(df)} Peptides] with less than [{peptide_num}] peptides.")
+        print(f"Removed [{len(set((remove_list)))} {df_type}] from [{df_original_len - len(df)} Peptides] with less than [{peptide_num}] peptides.")
 
         return df
 
@@ -819,7 +875,10 @@ class TaxaFuncAnalyzer:
             self.peptide_num_used['protein'] = 0
             sum_protein_params['quant_method'] = quant_method
             df_peptide_for_protein = self.detect_and_handle_outliers(df=self.original_df, **outlier_params)
-            self.protein_df = SumProteinIntensity(taxa_func_analyzer=self, df=df_peptide_for_protein).sum_protein_intensity( **sum_protein_params)
+            self.protein_df = SumProteinIntensity(taxa_func_analyzer=self, df=df_peptide_for_protein,
+                                                  peptide_num_threshold=sum_protein_params['peptide_num_threshold'],
+                                                  protein_separator = self.protein_separator
+                                                  ).sum_protein_intensity( **sum_protein_params)
             self.protein_df = self.data_preprocess(df=self.protein_df,df_name = 'protein', 
                                                    **data_preprocess_params)
             
@@ -1105,9 +1164,10 @@ if __name__ == '__main__':
                                                 'batch_meta': 'None', 
                                                 'processing_order': ['transform', 'normalize', 'batch']},
                     peptide_num_threshold = {'taxa': 3, 'func': 3, 'taxa_func': 3},
-                    keep_unknow_func=False, sum_protein=False, 
-                    sum_protein_params = {'method': 'razor', 'by_sample': False, 'rank_method': 'unique_counts', 'greedy_method': 'heap', 'peptide_num_threshold': 3},
-                    split_func=True, split_func_params = {'split_by': '|', 'share_intensity': False},
+                    keep_unknow_func=False,
+                    sum_protein=True, 
+                    sum_protein_params = {'method': 'anti-razor', 'by_sample': False, 'rank_method': 'unique_counts', 'greedy_method': 'heap', 'peptide_num_threshold': 3},
+                    split_func=False, split_func_params = {'split_by': '|', 'share_intensity': False},
                     taxa_and_func_only_from_otf=False, quant_method='sum'
                     )