+++ Docs/ChangeLog.md
@@ -1,7 +1,8 @@
 # Version: 1.115.4
 ## Date: 2024-10-07
 ### Changes:
-- TODO: use the peptide number for 'self.peptide_num_used' after filtering the minimum peptide number
+- Fix: Fixed the bug of when use Anydata moed, the report will raise error.
+- Change: changed the approche of filter the minimum number of peptides threshold for the protein.(Avaliable for Razor and Anti-Razor method)
 
 # Version: 1.115.3
 ## Date: 2024-10-04